Add third_party_login to user settings API (#103)
Currently, Oauth users are presented with a password field when resetting their password or when adding 2-factor authentication. For Oauth users, this doesn't make sense since they cannot have both a Oauth-linked account and a user/pass account. We need a way to tell them apart so that the app doesn't present them with the password field which doesn't do anything for them.

This adds a third_party_login boolean virtual field to the User struct which is hydrated when pulled from the database. This is set to true when the user has any oauth IDs saved in the database.

Refactored the tests a bit in Api.UserSettingsControllerTest to separate two-factor users, oauth users, and only set them up for the relevant tests.

This also creates the oauth_users table to catch up from our other system. This table already exists; we're only creating it here for tests.